app/requirements.txt
fastapi
uvicorn
pydub
requests
google-generativeai
gradio_client
